# backend/my-blog-hono/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "QBlog Hono API",
		"description": "RESTful API for QBlog - Blog and Portfolio Management System",
		"version": "1.0.0",
		"contact": {
			"name": "QBlog Support",
			"url": "https://qblog.dev"
		},
		"license": {
			"name": "MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:8787",
			"description": "Local Development Server"
		},
		{
			"url": "https://api.qblog.dev",
			"description": "Production Server"
		}
	],
...
	"paths": {
...
		"/api/projects": {
			"get": {
				"tags": ["Projects"],
				"summary": "List All Projects",
				"description": "Get all portfolio projects",
				"operationId": "listProjects",
				"responses": {
					"200": {
						"description": "Projects retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectsListResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/projects/{id}": {
			"get": {
				"tags": ["Projects"],
				"summary": "Get Project by ID",
				"description": "Retrieve a single project by numeric ID",
				"operationId": "getProjectById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "Project ID",
						"schema": {
							"type": "integer",
							"example": 1
						}
					}
				],
				"responses": {
					"200": {
						"description": "Project retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponse"
								}
							}
						}
					},
					"404": {
						"description": "Project not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/projects": {
			"get": {
				"tags": ["Projects"],
				"summary": "List Projects (Admin)",
				"description": "Get all projects with admin details (Admin only)",
				"operationId": "listProjectsAdmin",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Projects retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectsListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Projects"],
				"summary": "Create Project",
				"description": "Create a new project (Admin only)",
				"operationId": "createProject",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProjectRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Project created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing required fields",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/projects/{id}": {
			"put": {
				"tags": ["Projects"],
				"summary": "Update Project",
				"description": "Update a project by ID (Admin only)",
				"operationId": "updateProject",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "Project ID",
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProjectRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Project updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Project not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Projects"],
				"summary": "Delete Project",
				"description": "Delete a project by ID (Admin only)",
				"operationId": "deleteProject",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "Project ID",
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Project deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Project not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session_id"
			}
		},
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "error"
					},
					"error": {
						"type": "string",
						"example": "Error message"
					}
				}
			},
...
			"CreateProjectRequest": {
				"type": "object",
				"required": ["title", "description"],
				"properties": {
					"title": {
						"type": "string",
						"example": "My Awesome Project"
					},
					"description": {
						"type": "string",
						"example": "Description of the project"
					},
					"imageUrl": {
						"type": "string",
						"example": "https://example.com/image.jpg"
					},
					"technologies": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"example": ["React", "TypeScript", "Tailwind CSS"]
					},
					"githubLink": {
						"type": "string",
						"example": "https://github.com/user/project"
					},
					"liveDemoLink": {
						"type": "string",
						"example": "https://project.example.com"
					}
				}
			},
			"UpdateProjectRequest": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"example": "Updated Project Title"
					},
					"description": {
						"type": "string",
						"example": "Updated description"
					},
					"imageUrl": {
						"type": "string",
						"example": "https://example.com/new-image.jpg"
					},
					"technologies": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"example": ["Vue", "TypeScript"]
					},
					"githubLink": {
						"type": "string",
						"example": "https://github.com/user/updated-project"
					},
					"liveDemoLink": {
						"type": "string",
						"example": "https://updated-project.example.com"
					}
				}
			},
			"Project": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"title": {
						"type": "string",
						"example": "My Awesome Project"
					},
					"description": {
						"type": "string",
						"example": "Description of the project"
					},
					"imageUrl": {
						"type": "string",
						"example": "https://example.com/image.jpg"
					},
					"technologies": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"example": ["React", "TypeScript"]
					},
					"githubLink": {
						"type": "string",
						"example": "https://github.com/user/project"
					},
					"liveDemoLink": {
						"type": "string",
						"example": "https://project.example.com"
					},
					"dateAdded": {
						"type": "string",
						"format": "date-time",
						"example": "2025-10-28T00:00:00Z"
					}
				}
			},
			"ProjectsListResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"type": "object",
						"properties": {
							"projects": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Project"
								}
							}
						}
					}
				}
			},
			"ProjectResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"$ref": "#/components/schemas/Project"
					}
				}
			},
			"DeleteResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"type": "object",
						"properties": {
							"id": {
								"type": "string",
								"example": "slug_or_id"
							}
						}
					}
				}
			}
		}
	}
}